レビュー投稿機能実装

+++ app/controllers/books_controller.rb
@@ -1,12 +1,16 @@
 class BooksController < ApplicationController
+  before_action :authenticate_user!, only: [:new, :create, :edit, :update, :destroy]
 
   def index
-    @books = Book.all
-    @reviews = Review.all
+    @books = Book.all.order("created_at DESC").page(params[:page]).per(3)
+    @review = Review.all
+    @reviews = @review.includes(:book)
+    @user_reviews = current_user.reviews.order("created_at DESC") if current_user.present?
   end
 
   def new
     @book = Book.new
+    render  :_new
   end
 
   def create
@@ -14,33 +18,49 @@ class BooksController < ApplicationController
     if @book.save
       redirect_to root_path
     else
-      redirect_to  new_book_path
+      redirect_to new_book_path, flash: {notice: '本の保存に失敗しました'}
     end
   end
   
   def show
-    @review = Review.find(params[:id])
+    @book = Book.find(params[:id])
+    @review = Review.new
+    @reviews = @book.reviews.includes(:book)
   end
 
   def edit
     @book = Book.find(params[:id])
+    if @book.user_id == current_user.id
+      @book = Book.find(params[:id])
+    else
+      redirect_to action: 'show'
+    end
   end
 
   def update
     book = Book.find(params[:id])
     book.update(book_params)
-    redirect_to book_path(book.id)
+    if book.save
+      redirect_to book_path(book.id)
+    else
+      redirect_to action: 'update'
+    end
   end
 
   def destroy
     book = Book.find(params[:id])
-    book.destroy
-    redirect_to root_path
+    if book.user_id == current_user.id
+      book.destroy
+      redirect_to root_path
+    else
+      redirect_to action: 'show'
+    end
   end
 
   private
   def book_params
     params.require(:book).permit(:booktitle, :author, :bookimage).merge(user_id: current_user.id)
   end
+
 
 end
